examples/json-rpc/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
json-rpc-3==1.8.4
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
